New Package Struture
* Rename sub-modules
* Split out lookup into logical groupings
* Group computation commands into sub-folders
* Move Rtree fuctions from utils.perils to lookup.rtree
* Add support of older imports (Backwards compatible)
