+++ oasislmf/computation/hooks/post_analysis.py
@@ -1,0 +1,1 @@
+## TO DO -- placeholder for post analysis output modification hook
