+++ oasislmf/execution/bash.py
