+++ oasislmf/preparation/lookup.py
@@ -1,0 +1,19 @@
+__all__ = [
+    'OasisBuiltinBaseLookup',
+    'OasisBaseKeysLookup',
+    'OasisLookup',
+    'OasisPerilLookup',
+    'OasisVulnerabilityLookup',
+    'OasisLookupFactory'
+]
+
+
+## This is a backwards compatibility file for lookup code based on version 1.9.0 and older
+## Add deprecated warning, Load equiv classes from ..lookup. * 
+
+from ..lookup.factory import OasisLookupFactory
+from ..lookup.base import OasisBaseLookup as OasisBuiltinBaseLookup
+from ..lookup.interface import OasisLookupInterface as OasisBaseKeysLookup 
+from ..lookup.rtree import RTreeLookup as OasisLookup
+from ..lookup.rtree import RTreePerilLookup as OasisPerilLookup
+from ..lookup.rtree import RTreeVulnerabilityLookup as OasisVulnerabilityLookup
