+++ oasislmf/utils/defaults.py
@@ -17,6 +17,7 @@ __all__ = [
     'GROUP_ID_COLS',
     'API_EXAMPLE_AUTH',
     'KEY_NAME_TO_FILE_NAME',
+    'DEFAULT_RTREE_INDEX_PROPS',
     'KTOOLS_ALLOC_GUL_MAX',
     'KTOOLS_ALLOC_FM_MAX',
     'KTOOLS_FIFO_RELATIVE',
@@ -71,6 +72,34 @@ KEY_NAME_TO_FILE_NAME = {
     'profile_acc_json': 'profile_account.json',
     'profile_fm_agg_json': 'profile_fm_agg.json',
 }
+
+DEFAULT_RTREE_INDEX_PROPS = {
+    'buffering_capacity': 10,
+    'custom_storage_callbacks': None,
+    'custom_storage_callbacks_size': 0,
+    'dat_extension': 'dat',
+    'dimension': 2,
+    'filename': '',
+    'fill_factor': 0.7,
+    'idx_extension': 'idx',
+    'index_capacity': 100,
+    'index_id': None,
+    'leaf_capacity': 100,
+    'near_minimum_overlap_factor': 32,
+    'overwrite': True,
+    'pagesize': 4096,
+    'point_pool_capacity': 500,
+    'region_pool_capacity': 1000,
+    'reinsert_factor': 0.3,
+    'split_distribution_factor': 0.4,
+    'storage': 0,
+    'tight_mbr': True,
+    'tpr_horizon': 20.0,
+    'type': 0,
+    'variant': 2,
+    'writethrough': False
+}
+
 
 # Store index from merged source files (for later slice & dice)
 SOURCE_IDX = OrderedDict({
